💰 Add an org-global funding config
This updates eliminates the need to have the config copied into every
repository in the GitHub organization and allows for its centralized
maintenance.

This patch takes the config in the main repo as a base, extending it
with natively supported `buy_me_a_coffee`, `thanks_dev` as well as a
few custom links of a social nature.